+++ docs/CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   just the next one (e.g., both rhel-9.9 and rhel-9.10 when latest Z is rhel-9.8.z,
   but not rhel-9.7) (OSIDB-5535)
 
+### Added
+- Add On-demand API endpoint to trigger link_tracker_with_affects (OSIDB-5470)
+
 ## [5.18.0] - 2026-09-10
 ### Added
 - Add `resolved_dt` recording when a flaw enters the DONE workflow state (OSIDB-5467)